--- a/include/metall/detail/proc.hpp
+++ b/include/metall/detail/proc.hpp
@@ -10,9 +10,9 @@
 
 #ifdef _GNU_SOURCE
 #include <sys/sysinfo.h>
-#define SUPPORT_GET_CPU_CORE_NO true
+#define SUPPORT_GET_CPU_NO true
 #else
-#define SUPPORT_GET_CPU_CORE_NO false
+#define SUPPORT_GET_CPU_NO false
 #endif
 
 #include <thread>
@@ -21,8 +21,8 @@ namespace metall::mtlldetail {
 
 /// \brief Returns the number of the logical CPU core on which the calling
 /// thread is currently executing.
-inline int get_cpu_core_no() {
-#if SUPPORT_GET_CPU_CORE_NO
+inline int get_cpu_no() {
+#if SUPPORT_GET_CPU_NO
 
   const int cpu = ::sched_getcpu();
   if (cpu == -1) {
@@ -33,7 +33,7 @@ inline int get_cpu_core_no() {
 #else
 
 #ifdef METALL_VERBOSE_SYSTEM_SUPPORT_WARNING
-#warning "CPU core number is always 0"
+#warning "CPU number is always 0"
 #endif
   return 0;
 
@@ -41,8 +41,8 @@ inline int get_cpu_core_no() {
 }
 
 /// \brief Returns the number of the logical CPU cores on the system.
-inline int get_num_cpu_cores() {
-#if SUPPORT_GET_CPU_CORE_NO
+inline int get_num_cpus() {
+#if SUPPORT_GET_CPU_NO
   return get_nprocs_conf();
 #else
   return int(std::thread::hardware_concurrency());