This PR is for a completely new object cache implementation.

The new cache is:

- Uses much less memory
- Significantly faster initialization time, especially on a machine that has many CPU cores
- More compact and packed internal data structure
- Yet, achieves the same Metall performance on the dynamic graph construction workload

Simple create and close test results (just creating and closing Metall manager).

- New ver
Create 0.0440739s
Close 0.00165344s

- Old ver
Open 0.658001s
Close 0.116477s